refactor(mcp): index.ts から marker + actionGroup handlers を分離
## 背景

designer-mcp/src/index.ts が 1350 行の巨大 switch statement になっており、
tool category ごとの責務が混在していた。#302 の後続作業として、
tool handler を category 別モジュールに分離。

## 変更

- `src/mcpHelpers.ts` 新設: saveAndBroadcast + ToolHandler 型定義
- `src/handlers/marker.ts` 新設: list/find_all/add/resolve/remove_marker (5 tool)
- `src/handlers/actionGroup.ts` 新設: list/get/add/update/delete_action_group,
  add_action/add_step/update_step/remove_step/move_step,
  set_maturity/add_step_note/add_catalog_entry/remove_catalog_entry (14 tool)
- `src/index.ts`: dispatcher pattern で各 handler に委譲、該当無ければ
  既存 switch に fallback

## 今後の余地

残り 35 ケース (canvas + screen + custom blocks + table + export + tab + save)
は未分離。今回の pattern に従って同じく別モジュール化可能。

## 検証

- designer-mcp tsc: pass
- designer-mcp vitest: 37/37 pass (無回帰)
- index.ts: 1268 → 1037 lines (-231)
